+++ packages/terra-form-input/CHANGELOG.md
@@ -3,15 +3,17 @@ ChangeLog
 
 Unreleased
 ----------
+### Removed
+* Removed explicit calls to set `caret-color` to `transparent` in WDIO tests
 
 2.19.0 - (July 17, 2019)
 ------------------
-## Changed
+### Changed
 * Revert replaced `Field.defaultProps.errorIcon` with `<IconError>` component.
 
 2.18.0 - (July 17, 2019)
 ------------------
-## Changed
+### Changed
 * Replaced `Field.defaultProps.errorIcon` with `<IconError>` component.
 
 2.17.0 - (July 16, 2019)