+++ packages/terra-form-select/CHANGELOG.md
@@ -3,6 +3,8 @@ ChangeLog
 
 Unreleased
 ----------
+### Removed
+* Removed explicit calls to set `caret-color` to `transparent` in WDIO tests
 
 5.23.0 - (July 17, 2019)
 ------------------