+++ packages/terra-form-textarea/CHANGELOG.md
@@ -3,6 +3,8 @@ ChangeLog
 
 Unreleased
 ----------
+### Removed
+* Removed explicit calls to set `caret-color` to `transparent` in WDIO tests
 
 3.19.0 - (July 17, 2019)
 ------------------